refactor(spdx-utils): Move tests out of `SpdxExpressionParserTest`
These tests do not test the SPDX parser in the grammar sense, but the
`parse()` function of the `SpdxExpression` class. As such, move these
tests to `SpdxExpressionTest`. This also avoids confusion with

    utils/spdx/src/test/kotlin/parser/SpdxExpressionParserTest.kt

of the same name.
